+++ client/src/pages/Ask.jsx
@@ -1,1 +1,6 @@
 // 질문하는 페이지
+const Ask = () => {
+  return <div>Ask 페이지 입니다</div>;
+};
+
+export default Ask;
