+++ client/src/pages/Main.jsx
@@ -1,0 +1,5 @@
+const Main = () => {
+  return <div>Main 페이지 입니다</div>;
+};
+
+export default Main;
