+++ client/src/pages/MyProfile.jsx
@@ -1,0 +1,5 @@
+const MyProfile = () => {
+  return <div>MyProfile 페이지 입니다</div>;
+};
+
+export default MyProfile;
