+++ client/src/pages/Question.jsx
@@ -1,1 +1,6 @@
 // 질문 보는 페이지
+const Question = () => {
+  return <div>Question 페이지 입니다</div>;
+};
+
+export default Question;
